requirements.txt
pywebview==4.4.1
PyYAML==6.0.2
pywin32==306
mss==9.0.1
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==2.0.1
pyautogui==0.9.54
winrt==1.0.21033.1
